Refactor names and delete unused files
